Factor out yes-no-radios-form-field.tsx.
