+++ frontend/lib/pages/onboarding-step-3.tsx
@@ -6,7 +6,8 @@ import autobind from 'autobind-decorator';
 import { Link, Route } from 'react-router-dom';
 import { NextButton, BackButton } from "../buttons";
 import { IconLink } from "../icon-link";
-import { RadiosFormField, YesNoRadiosFormField } from '../form-fields';
+import { RadiosFormField } from '../form-fields';
+import { YesNoRadiosFormField } from "../yes-no-radios-form-field";
 import { ReactDjangoChoices } from '../common-data';
 import { OnboardingStep3Mutation } from '../queries/OnboardingStep3Mutation';
 import { Modal, BackOrUpOneDirLevel } from '../modal';
