+++ frontend/lib/tests/pages/fee-waiver.tsx
@@ -5,7 +5,8 @@ import Page from "../../page";
 import { SessionUpdatingFormSubmitter } from '../../forms';
 import { FeeWaiverMutation } from '../../queries/FeeWaiverMutation';
 import Routes from '../../routes';
-import { RadiosFormField, CheckboxFormField, YesNoRadiosFormField } from '../../form-fields';
+import { RadiosFormField, CheckboxFormField } from '../../form-fields';
+import { YesNoRadiosFormField } from "../../yes-no-radios-form-field";
 import { toDjangoChoices } from '../../common-data';
 import { IncomeFrequencyChoices, getIncomeFrequencyChoiceLabels } from '../../../../common-data/income-frequency-choices';
 import { BackButton, NextButton } from '../../buttons';
